Changes to nvme_control_common .c and .h
