+++ src/control/lib/spdk/include/nvme_control_common.h
@@ -26,7 +26,8 @@
 
 struct ctrlr_entry {
 	struct spdk_nvme_ctrlr	*ctrlr;
-	const char		*tr_addr;
+	struct spdk_pci_addr	pci_addr;
+	int 			socket_id;
 	struct ctrlr_entry	*next;
 };
